// comonl-comonlwcl/src/app/modules/comonlapi/model/apiError.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ApiError { 
    code?: string;
    group?: string;
    errorMessage?: string;
    params?: { [key: string]: any; };
    groupParams?: { [key: string]: any; };
    type?: string;
    fullErrorMessage?: string;
}